==> 8_pseudosync_Moulder.Rmd <==
---
title: "Pseudosynchronisation for ML speech"
author: "I. S. Plank"
date: "03/01/2023"
output: html_document
---

```{r setup, include=FALSE}

knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = '/home/emba/Documents/ML_speech/Data/')

library(ggpubr)
library(ggsignif)
library(tidyverse)
library(knitr)       # kable
library(BayesFactor) # ttestBF
library(rMEA)
library(pastecs)
source("/home/emba/Documents/ML_speech/MLSPE_scripts/fakeMEA_function.R")

# how many iterations for shuffling?
n = 100

lb = "#56B4E9"
db = "#0072B2"
gr = "#009E73"

```

<style type="text/css">
.main-container {
  max-width: 1100px;
  margin-left: auto;
  margin-right: auto;
}
</style>

# General information

This R Markdown document analyses pseudosynchronisation in the speech data of the ML study. To run this script, one needs the following file: 

* "ML_turns.csv" which contains all turn information needed for section sliding
* "ML_indi.csv" which contains the actual synchronisation values for comparison
* "ML_dyad.csv" which contains the actual synchrony values for comparison

## Software versions

```{r version}
R.Version()$version.string
packageVersion('ggpubr')
packageVersion('ggsignif')
packageVersion('tidyverse')
packageVersion('knitr')
packageVersion('BayesFactor')
packageVersion('rMEA')
packageVersion('pastecs')
```

# Individual level: turn-based pseudosynchronisation

For turn-based pseudosynchronisation, this script uses an approach described in Moulder et al. (2018, *Psychol Methods*) to test the following hypothesis: "Synchrony does not exist between sections of size *m* in these two time series" (section sliding). This segment shuffling, from Moulder et al., "requires researchers to cut a time series *X* into shorter sections of size *m* which are randomly appended to one another to create a new time series *Xs* until no section is in its original position". This new time series is then paired with an unchanged time series *Y*. Since the synchronisation is already based on sections  or speech turns in the conversation, no data shuffling will be performed. Since each dyad can have a unique amount of turns, no dyad shuffling will be performed. 

For each speaker in each task, the order of the previous turns are shuffled `r n` times while the order of the current turns remain unchanged. Note that no check for uniqueness is performed. In the end, an average for this specific speaker and task is calculated from the `r n` pseudosynchrony values. 

## Segment shuffling: is there synchrony between turns in these two time series?

```{r indi_turn}

filename = "ML_indi_st-shuffle.RData"

if (!file.exists(filename)) {
  
  speaker  = c()
  psyncint = c()
  psyncpit = c()
  psyncart = c()
  
  df = read_csv("ML_turns.csv", show_col_types = F) %>% 
    mutate(speaker = paste(name, speaker, sep = "_")) %>%
    select(c(speaker, pit_turn, pit_prev_turn, 
             int_turn, int_prev_turn, 
             art_prev_turn, art_turn))

  for (s in unique(df$speaker)) {
    speaker = c(speaker, rep(s, times = n))
    df.sel = df %>% filter(speaker == s)
    for (i in 1:n) {
      # shuffling 
      df.sel$pit_prev  = sample(df.sel$pit_prev_turn)
      df.sel$int_prev  = sample(df.sel$int_prev_turn)
      df.sel$art_prev = sample(df.sel$art_prev_turn)
      psyncint = c(psyncint, abs(cor(df.sel$int_turn,df.sel$int_prev, use = "na.or.complete")))
      psyncpit = c(psyncpit, abs(cor(df.sel$pit_turn,df.sel$pit_prev, use = "na.or.complete")))
      psyncart = c(psyncart, abs(cor(df.sel$art_turn,df.sel$art_prev, use = "na.or.complete")))
    }
  }
  df.s_shuffle = data.frame(speaker,psyncint, psyncpit, psyncart)
  save(df.s_shuffle,file=filename)
} else {
  load(filename)
}

```

```{r indi_stats}

df.sync = read_csv("ML_indi.csv", show_col_types = F) %>% 
  separate(
    subject, c("ML", "dyad", "speaker")
  ) %>%
  mutate(
    speaker = paste(ML, dyad, task, speaker, sep = "_")
  ) %>%
  select(c(speaker, pit_sync, int_sync, art_sync))

df.agg  = df.s_shuffle %>% group_by(speaker) %>%
  summarise(
    pit_psync  = mean(psyncpit),
    int_psync  = mean(psyncint),
    art_psync  = mean(psyncart)
  )

df.agg = merge(df.sync,df.agg)

kable(stat.desc(df.agg %>% select(where(is.numeric)) %>% na.omit()))

res = ttestBF(x = df.agg[!is.na(df.agg$pit_sync),]$pit_sync,
        y = df.agg[!is.na(df.agg$pit_sync),]$pit_psync,
        paired = T) 
res
res@bayesFactor[["bf"]]
res = ttestBF(x = df.agg[!is.na(df.agg$int_sync),]$int_sync,
        y = df.agg[!is.na(df.agg$int_sync),]$int_psync,
        paired = T) 
res
res@bayesFactor[["bf"]]
res = ttestBF(x = df.agg[!is.na(df.agg$art_sync),]$art_sync,
        y = df.agg[!is.na(df.agg$art_sync),]$art_psync,
        paired = T)  
res
res@bayesFactor[["bf"]]

```

The t-tests show that there is a significant difference between the pseudosynchronisation values obtained through segment shuffling and the real synchronisation values in the case of turn-based synchronisation of intensity, pitch and articulation rate, with the real synchronisation values being higher. 

```{r plot_indi, echo=FALSE, eval=T, fig.width = 12, fig.align = "center", warning = F, message = F}

df.pit = pivot_longer(df.agg, cols = c(pit_sync, pit_psync))
df.int = pivot_longer(df.agg, cols = c(int_sync, int_psync))
df.art = pivot_longer(df.agg, cols = c(art_sync, art_psync))

p1 = ggplot(df.pit, aes(x = name, y = value)) + geom_boxplot(aes(fill = name)) + ggtitle("Pitch") + 
  theme_minimal() + ylab("") + xlab("")  + scale_fill_manual(values = c(gr,db)) + theme(legend.position = "none") + 
  geom_signif(comparisons = list(c("pit_sync", "pit_psync")), map_signif_level = c("*"=0.05),
              test = "t.test")

p2 = ggplot(df.int, aes(x = name, y = value)) + geom_boxplot(aes(fill = name)) + 
  ggtitle("Intensity") + 
  theme_minimal() + ylab("") + xlab("")  + scale_fill_manual(values = c(gr,db)) + theme(legend.position = "none") +
  geom_signif(comparisons = list(c("int_sync", "int_psync")), map_signif_level = c("*"=0.05),
              test = "t.test")

p3 = ggplot(df.art, aes(x = name, y = value)) + geom_boxplot(aes(fill = name)) + 
  ggtitle("Articulation rate") + 
  theme_minimal() + ylab("") + xlab("")  + scale_fill_manual(values = c(gr,db)) + theme(legend.position = "none") +
  geom_signif(comparisons = list(c("art_sync", "art_psync")), map_signif_level = c("*"=0.05),
              test = "t.test")

ggarrange(p1, p2, p3, ncol = 3, nrow = 1)

```

# Dyad level: timecourse pseudosynchrony based on CCF

We again use segment shuffling to compare pseudosynchrony and synchrony values on the dyad level. Pitch and intensity synchrony was calculated using the CCF function from rMEA. Therefore, we use the same function with the same window size but with shuffled windows for calculating pseudosynchrony. 

For each speaker in each task, their time series is shuffled `r n/2` times and then paired with the unchanged time series of their partner in the dyad. Note that no check for uniqueness is performed. In the end, an average for this specific dyad and task is calculated from the `r n` pseudosynchrony values. 

## Segment shuffling: is there synchrony between sections of 16 seconds (pitch, intensity) in these two time series?

```{r dyad-s}

filename = "ML_dyad_s-shuffle.RData"

if (!file.exists(filename)) {
  
  # create new data frame for pseudosync values
  cols = c("dyad", "dyad type", "task", "pit_psync_MEA", "int_psync_MEA")
  df.psync = data.frame(matrix(nrow = 0, ncol = length(cols)))
  colnames(df.psync) = cols
  
  # load actual sync values
  df.dyad = read_csv("ML_dyad.csv", show_col_types = F)
  
  sr = 100
  
  # go through all dyads
  for (i in 1:nrow(df.dyad)) {
    # load the respective data of continuous pitch and intensity 
    ptn = paste("^ch_.*", df.dyad$dyad[i], ".*", df.dyad$task[i], ".*_cont.csv", sep = "")
    fls = list.files(pattern = ptn)
    df_L = read_delim(fls[1], delim = ";", show_col_types = F, col_types = cols(.default = "d")) 
    df_R = read_delim(fls[2], delim = ";", show_col_types = F, col_types = cols(.default = "d")) 
    # resample to 100Hz
    L_pit = df_L$pitch
    L_pit[is.na(L_pit)] = 0
    L_pit = signal::resample(L_pit,1,1000/sr) 
    L_int = df_L$int
    L_int[is.na(L_int)] = 0
    L_int = signal::resample(L_int,1,1000/sr) 
    R_pit = df_R$pitch
    R_pit[is.na(R_pit)] = 0
    R_pit = signal::resample(R_pit,1,1000/sr) 
    R_int = df_R$int
    R_int[is.na(R_int)] = 0
    R_int = signal::resample(R_int,1,1000/sr) 
    # create fakeMEA and divide into segments
    mea_int  = fakeMEA(L_int,R_int,sr,"int_L","int_R")
    mea_pit  = fakeMEA(L_pit,R_pit,sr,"pit_L","pit_R")
    mea_int_L = split(mea_int[["all_01_01"]][["MEA"]][["s1Name"]], floor(seq_along(mea_int[["all_01_01"]][["MEA"]][["s1Name"]])/16))
    mea_int_R = split(mea_int[["all_01_01"]][["MEA"]][["s2Name"]], floor(seq_along(mea_int[["all_01_01"]][["MEA"]][["s2Name"]])/16))
    mea_pit_L = split(mea_pit[["all_01_01"]][["MEA"]][["s2Name"]], floor(seq_along(mea_pit[["all_01_01"]][["MEA"]][["s1Name"]])/16))
    mea_pit_R = split(mea_pit[["all_01_01"]][["MEA"]][["s2Name"]], floor(seq_along(mea_pit[["all_01_01"]][["MEA"]][["s2Name"]])/16))
    # shuffle each side n/2 times
    for (j in 1:(n/2)) {
      
      # shuffling left
      mea_int[["all_01_01"]][["MEA"]][,2] = unlist(mea_int_R)
      mea_int[["all_01_01"]][["MEA"]][,1] = unlist(sample(mea_int_L))
      mea_int = MEAccf(mea_int,2,16,8)
      mea_pit[["all_01_01"]][["MEA"]][,2] = unlist(mea_pit_R)
      mea_pit[["all_01_01"]][["MEA"]][,1] = unlist(sample(mea_pit_L))
      mea_pit = MEAccf(mea_pit,2,16,8)
      # create new row with mean of peaks of windows
      new = c(df.dyad$dyad[i],
              df.dyad$`dyad type`[i],
              df.dyad$task[i],
              mean(apply(mea_int$all_01_01$ccf,1,max)),
              mean(apply(mea_pit$all_01_01$ccf,1,max)))
      df.psync[nrow(df.psync) + 1,] = new      
      
      # shuffling right
      mea_int[["all_01_01"]][["MEA"]][,2] = unlist(sample(mea_int_R))
      mea_int[["all_01_01"]][["MEA"]][,1] = unlist(mea_int_L)
      mea_int = MEAccf(mea_int,2,16,8)
      mea_pit[["all_01_01"]][["MEA"]][,2] = unlist(sample(mea_pit_R))
      mea_pit[["all_01_01"]][["MEA"]][,1] = unlist(mea_pit_L)
      mea_pit = MEAccf(mea_pit,2,16,8)
      # create new row with mean of peaks of windows
      new = c(df.dyad$dyad[i],
              df.dyad$`dyad type`[i],
              df.dyad$task[i],
              mean(apply(mea_int$all_01_01$ccf,1,max)),
              mean(apply(mea_pit$all_01_01$ccf,1,max)))
      df.psync[nrow(df.psync) + 1,] = new
    }
  }
  # save everything
  save(df.psync,file=filename)
} else {
  load(filename)
}

```

```{r dyad-s_stats, message=F}

df.sync = read_csv("ML_dyad.csv", show_col_types = F) %>% 
  select(c(dyad, `dyad type`, task, pit_sync_MEA, int_sync_MEA))

df.agg  = df.psync %>% 
  mutate(
    pit_psync_MEA = as.numeric(pit_psync_MEA),
    int_psync_MEA = as.numeric(int_psync_MEA)
  ) %>%
  group_by(dyad, task) %>%
  summarise(
    pit_psync_MEA  = mean(pit_psync_MEA),
    int_psync_MEA  = mean(int_psync_MEA)
  )

df.agg = merge(df.sync,df.agg) 

kable(stat.desc(df.agg %>% select(where(is.numeric)) %>% na.omit()))

res = ttestBF(x = df.agg[!is.na(df.agg$pit_sync_MEA),]$pit_sync_MEA,
        y = df.agg[!is.na(df.agg$pit_sync_MEA),]$pit_psync_MEA,
        paired = T) 
res
res@bayesFactor[["bf"]]
res = ttestBF(x = df.agg[!is.na(df.agg$int_sync_MEA),]$int_sync_MEA,
        y = df.agg[!is.na(df.agg$int_sync_MEA),]$int_psync_MEA,
        paired = T) 
res
res@bayesFactor[["bf"]] 

```

The t-tests show that there is a significant difference between the pseudosynchrony values obtained through segment shuffling and the real synchrony values in the case of MEA-based CCF synchrony of intensity and pitch, with the real synchrony values being higher.

```{r plot_dyad-s, echo=FALSE, eval=T, fig.width = 12, fig.align = "center", warning = F, message = F}

df.pit = pivot_longer(df.agg, cols = c(pit_sync_MEA, pit_psync_MEA))
df.int = pivot_longer(df.agg, cols = c(int_sync_MEA, int_psync_MEA))

p1 = ggplot(df.pit, aes(x = name, y = value)) + geom_boxplot(aes(fill = name)) + 
  ggtitle("Pitch") + 
  theme_minimal() + ylab("") + xlab("")  + scale_fill_manual(values = c(gr,db)) + theme(legend.position = "none") +
  geom_signif(comparisons = list(c("pit_sync_MEA", "pit_psync_MEA")), map_signif_level = c("*"=0.05),
              test = "t.test")

p2 = ggplot(df.int, aes(x = name, y = value)) + geom_boxplot(aes(fill = name)) + 
  ggtitle("Intensity") + 
  theme_minimal() + ylab("") + xlab("")  + scale_fill_manual(values = c(gr,db)) + theme(legend.position = "none") +
  geom_signif(comparisons = list(c("int_sync_MEA", "int_psync_MEA")), map_signif_level = c("*"=0.05),
              test = "t.test")

ggarrange(p1, p2, ncol = 2, nrow = 1)

```
